libavcodec/hevc_mp4toannexb_bsf: copy extradata to the beginning of an IRAP frame
It is possible that an IRAP frame in input AVPacket contains VPS, SPS
and PPS, and these headers should take effect. However the prepended
extradata might override these headers. This patch copies extradata to
the beginning of an IRAP frame, hence VPS, SPS and PPS from the input
AVPacket will take effect if they are present.

This should fix #7799

Signed-off-by: Haihao Xiang <haihao.xiang@intel.com>